// src/app/shared/models/label-model.ts
import { LabelData } from 'desy-angular';

export class LabelModel implements LabelData{
  ariaControls: string;
  ariaCurrent: string;
  ariaDescribedBy: string;
  ariaDisabled: string;
  ariaErrorMessage: string;
  ariaExpanded: string;
  ariaHasPopup: string;
  ariaHidden: string;
  ariaLabel: string;
  ariaLabelledBy: string;
  ariaLive: string;
  classes: string;
  for: string;
  html: string;
  id: string;
  isPageHeading: boolean;
  role: string;
  text: string;
  tabindex: string;
}